set ValidFor (Valid-Until header in Release files)
reason: The Valid-Until header allows APT clients to find out when
active network attackers are holding back newer Release files.

Why 50 days: 2 * 3-weeks iterations (allows to skip one release
without action) + some margin

If we skip 2 releases or more, we should re-generate the Release files
(reprepro export) to bump the Valid-Until header so that the clients
can see that the non-release was real.

I did not modify phoenix yet, so that we can see if this causes
trouble on wazo-dev first.